fix: address PR review — TX wrap, 30s window, reader lock, reset
- Change totalTxFrames to uint64_t (prevent 32-bit wrap)
- Handle TX counter wrap in snapshot delta calculation
- Use clock_gettime_nsec_np for 30s RX ID window (not max RX timestamp)
- Lock readers_lock in rxReaderCount() (prevent data race)
- Reset secTx and lastTxCount in reset()

+++ ican/perf/DashboardMetricsEngine.hpp
@@ -31,7 +31,7 @@ struct DashboardSnapshot {
     // Cumulative metrics
     uint64_t totalMessages;      // RX frames received
     uint64_t totalBytes;         // RX bytes
-    uint32_t totalTxFrames;      // TX frames written (app-side)
+    uint64_t totalTxFrames;      // TX frames written (app-side)
     int      rxReaderCount;      // active RX reader clients
     int      txWriterCount;      // active TX writer clients
     int      rxUniqueIds30s;    // distinct RX CAN IDs in last 30s
